requirements.txt
numpy==1.23.1
opencv-python==4.8.1.78
flask==2.1.2
werkzeug==2.1.2
jinja2==3.1.2
gunicorn==20.1.0
pillow==9.2.0
# gevent>=22.10.2,<23.0.0
boto3==1.24.36